requirements.txt
twikit>=2.3.0
anthropic>=0.80.0
streamlit>=1.40.0
python-dotenv>=1.0.0
